Create CreateCSR.PS1
This script prompts the user to enter the required information for the certificate, such as the certificate name, country code, state or province name, locality or city name, organization name, organizational unit name, and common name. The script then creates a new X509Certificate2 object and sets the subject, key algorithm, key length, friendly name, and signature algorithm properties. Finally, the certificate request is exported to a file with the .csr extension, using the certificate name entered by the user.